simplify chat routing
